API Platform/fastapi/requirements.txt
fastapi[standard]
pydantic-settings
uvicorn[standard]
requests